syscall.c: use `dup3` in place of `getuid`
Use the `dup3` syscall in place of the `getuid` syscall in order to
avoid unpredictability of the return value. Specifically, in the CI
(which runs within a Docker container) the user ID is 0, while on a
developer machine it's typically 1000. On the other hand, with `dup3` we
can choose the expected return value predictably.

The `dup3` is used in place of `dup2` since `aarch64` does not expose
`dup2`.

This commit also drops the special handling of MIPS, which is likely no
longer necessary.