linked_list.c: avoid calling malloc
This commit removes calls to malloc from the linked_list.c test.
This is intended, for two reasons:
- to prevent the compiler from trying to be smart around those calls,
  such as eliding heap usage
- because the DataLayoutAnalysis of revng-c does not really handle
  mallocs gracefully

Instead of using malloc, the code tricks the compiler using char * in
argv as if they were pointers to data structures of the correct type.
Obviously the compiled program is broken, but the generated assembly is
perfectly fine to be analyzed for lifting and decompilation.