Revamp decompilation tests
This commit removes old obsolete decompilation tests, that were no
longer used, and replaces them with tests for various features of the
decompiler.
* DLA capability to recover complex data structure like linked-lists
  and arrays.
* DLA capability to update segment and section types, so that we emit
  nice looking accesses to segments in C.
* Capability to emit nice looking integer literals in C
* Capability to emit inline string literals in C and update the model
  types of the segments containg such string literals.

All the actual tests are implmented in revng-c, based on FileCheck.
Here we're adding only the original source, and recipes to compile the
code to generated the binaries that are then used in revng-c for tests.